app/requirements.txt
requests
Flask
beautifulsoup4
python-dotenv
lxml-html-clean
pymysql
gunicorn
google-cloud-secret-manager